controllerやview周りのリファクタ
